Added Enterprise Implementation

+++ AE.Domain/Abstract/AuditableEntity.cs
@@ -7,9 +7,9 @@ namespace AE.Domain.Abstract
     public abstract class AuditableEntity: BaseEntity
     {
         public DateTime CreatedOn {get; set; }
-        public int CreatedBy { get; set; }
+        public long CreatedBy { get; set; }
         public DateTime ModifiedOn {get; set; }
-        public int ModifiedBy { get; set; }
+        public long ModifiedBy { get; set; }
 
     }
 }
